Rename BasePairsRuler to FeatureLengthRuler
