+++ src/shared/components/feature-length-ruler/FeatureLengthRuler.scss
