+++ src/shared/components/feature-length-ruler/FeatureLengthRuler.tsx
@@ -32,11 +32,11 @@ It follows the following rules for displaying labelled and unlabelled ticks
 import React, { useEffect } from 'react';
 import { scaleLinear, ScaleLinear } from 'd3';
 
-import { getTicks } from './basePairsRulerHelper';
+import { getTicks } from './featureLengthRulerHelper';
 
 import { getCommaSeparatedNumber } from 'src/shared/helpers/formatters/numberFormatter';
 
-import styles from './BasePairsRuler.scss';
+import styles from './FeatureLengthRuler.scss';
 
 type Ticks = {
   ticks: number[];
@@ -52,10 +52,10 @@ type Props = {
   width: number; // number of pixels allotted to the axis on the screen
   unitsLabel?: string; // optional label showing what the ruler is measuring; will be displayed at the start of the ruler if present
   onTicksCalculated?: (payload: TicksAndScale) => void; // way to pass the ticks to the parent if it is interested in them
-  standalone: boolean; // wrap the component in an svg element if true
+  standalone?: boolean; // wrap the component in an svg element if true
 };
 
-const FeatureLengthAxis = (props: Props) => {
+const FeatureLengthRuler = (props: Props) => {
   const domain = [1, props.length];
   const range = [0, props.width];
   const scale = scaleLinear().domain(domain).rangeRound(range);
@@ -113,8 +113,4 @@ const FeatureLengthAxis = (props: Props) => {
   );
 };
 
-FeatureLengthAxis.defaultProps = {
-  standalone: false
-};
-
-export default FeatureLengthAxis;
+export default FeatureLengthRuler;
